+++ instrumentation/opentelemetry-instrumentation-wsgi/src/opentelemetry/instrumentation/wsgi/__init__.py
@@ -222,9 +222,9 @@ from opentelemetry.instrumentation._semconv import (
     _OpenTelemetryStabilitySignalType,
     _report_new,
     _report_old,
-    _server_active_requests_count_attrs_new,
+    _server_active_requests_count_attrs_new_effective,
     _server_active_requests_count_attrs_old,
-    _server_duration_attrs_new,
+    _server_duration_attrs_new_effective,
     _server_duration_attrs_old,
     _set_http_flavor_version,
     _set_http_method,
@@ -454,7 +454,7 @@ def _parse_active_request_count_attrs(
     return _filter_semconv_active_request_count_attr(
         req_attrs,
         _server_active_requests_count_attrs_old,
-        _server_active_requests_count_attrs_new,
+        _server_active_requests_count_attrs_new_effective(),
         sem_conv_opt_in_mode,
     )
 
@@ -465,7 +465,7 @@ def _parse_duration_attrs(
     return _filter_semconv_duration_attrs(
         req_attrs,
         _server_duration_attrs_old,
-        _server_duration_attrs_new,
+        _server_duration_attrs_new_effective(),
         sem_conv_opt_in_mode,
     )