+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Add `OTEL_PYTHON_HTTP_SERVER_REQUEST_DURATION_SERVER_ATTRIBUTES_ENABLED` flag to add 
+  `server.address` and `server.port` for `http.server.request.duration`
+  ([#2597](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/2597))
+- Add `OTEL_PYTHON_HTTP_SERVER_ACTIVE_REQUESTS_COUNT_SERVER_ATTRIBUTES_ENABLED` flag to add 
+  `server.address` and `server.port` for `http.server.active_requests`
+  ([#2597](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/2597))
 - `opentelemetry-instrumentation-fastapi` Add autoinstrumentation mechanism tests.
   ([#2860](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/2860))